ekf2: move GNSS hgt control to file
